Rewrite specs with test substitutes rather than mocks
The mocks OK, in that that they did the job. But they were also
confusing. Especially when testing an object which itself is primarily a
proxy object. It was often unclear if the object under test was doing
the right thing, or if the magic of RSpec doubles were making it seem
like things were OK.

Anyhow, it's easy enough to, mostly, save for some particularly awkward
APIs, use a hand-rolled test substitute. And that's what this does.